Cast to double to avoid overflow.

+++ src/usr.sbin/ntpd/util.c
@@ -45,13 +45,13 @@ gettime(void)
 	if (gettimeofday(&tv, NULL) == -1)
 		fatal("gettimeofday");
 
-	return (tv.tv_sec + JAN_1970 + 1.0e-6 * tv.tv_usec);
+	return ((double)tv.tv_sec + JAN_1970 + 1.0e-6 * tv.tv_usec);
 }
 
 double
 gettime_from_timeval(struct timeval *tv)
 {
-	return (tv->tv_sec + JAN_1970 + 1.0e-6 * tv->tv_usec);
+	return ((double)tv->tv_sec + JAN_1970 + 1.0e-6 * tv->tv_usec);
 }
 
 time_t
